# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.9.0"
  },
  "paths": {
...
    "/sampling/approve": {
      "post": {
        "tags": [
          "super::routes::sampling"
        ],
        "operationId": "handle_sampling_approval",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sampling approval processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SamplingApprovalRequest": {
        "type": "object",
        "required": [
          "session_id",
          "id",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "The user's action: approve, deny, or edit"
          },
          "edited_messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SamplingMessage"
            },
            "description": "If action is \"edit\", this contains the edited messages",
            "nullable": true
          },
          "id": {
            "type": "string",
            "description": "The sampling confirmation request ID"
          },
          "session_id": {
            "type": "string",
            "description": "The session ID for the current agent session"
          }
        }
      },
...
      "SamplingMessage": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "description": "The actual content of the message"
          },
          "role": {
            "type": "string",
            "description": "The role of the message sender (User or Assistant)"
          }
        }
      },
      "SamplingResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Optional message",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Status of the response"
          }
        }
      },
...
    }
  }
}